requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
DateTime==4.3
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
html5lib==1.0.1
Jinja2==2.10
lxml==4.2.1
Pillow==5.3.0
ptyprocess==0.5.2
PyMySQL==0.9.3
pytesseract==0.2.5
requests==2.18.4
SQLAlchemy==1.2.15
tesseract==0.1.3


